requirements.txt
coverage >= 5.0.3
hydra-core ~= 1.3
lark <= 1.1.8
pre-commit
pytest ~= 7.4
pytorch-lightning >= 0.7.6
tensorflow >= 2.0.0
tensorrt >= 7.0.0
torch >= 2.1.0
torch_tb_profiler >= 0.2.0
torchvision >= 0.4.0
